requirements.txt
db-dtypes
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
numpy
openpyxl
python-dotenv
jupyter
notebook
ipykernel
seaborn
ipywidgets
scikit-learn
xgboost
pydantic
fastapi
uvicorn[standard]
pyarrow